% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlluvialAE.R
\name{AlluvialAE}
\alias{AlluvialAE}
\title{AlluvialAE}
\usage{
AlluvialAE(data, N, choixEI, vect_grade, listcol, unit)
}
\arguments{
\item{data}{dataframe}

\item{N}{N}

\item{choixEI}{choixEI}

\item{vect_grade}{vect_grade}

\item{listcol}{listcol}

\item{unit}{unit}
}
\value{
A plot
}
\description{
helper for PanelLasagnaAlluvialAE
}
