% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarChart.R
\name{BarChart}
\alias{BarChart}
\title{BarChart}
\usage{
BarChart(data, group, Type, choixEI, vect_grade, listcol)
}
\arguments{
\item{data}{A dataframe}

\item{group}{A groupe}

\item{Type}{A type}

\item{choixEI}{choixEI}

\item{vect_grade}{vect_grade}

\item{listcol}{listcol}
}
\value{
A plot
}
\description{
helper for PanelLasagnaAlluvialAE
}
