% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButterflyBarChartSup.R
\name{ButterflyBarChartSup}
\alias{ButterflyBarChartSup}
\title{ButterflyBarChartSup}
\usage{
ButterflyBarChartSup(
  baseEI,
  baseTr,
  idvar,
  SOCvar,
  gradevar,
  SAEvar = NULL,
  TTTYN = NULL,
  ARMvar,
  varsup = 3,
  ARMe = NULL,
  tri = "RDAll",
  trivar = NULL,
  listcol = c("red", "deepskyblue2")
)
}
\arguments{
\item{baseEI}{la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item la SOC correspondante (SOC_COD) : label (et non code) au format character obligatoire et sans manquants
   \item le grade de l’EI (gradevar) : au format numeric obligatoire en partant de 1 (et allant jusqu’au grade max par exemple 5) et sans manquants
   \item si l’EI est de type sérieux (SAEvar) : codé en “Yes”/“No” obligatoire
}}

\item{baseTr}{la base des traitements avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item [non obligatoire] une variable pour indiquer si le traitement à été pris (sert à détecter les patients “à risque”) TTTYN on recherchera les occurences en “Yes” donc doit être codée en “Yes”/“No” en character
}}

\item{idvar}{column name}

\item{SOCvar}{column name}

\item{gradevar}{column name}

\item{SAEvar}{column name}

\item{TTTYN}{column name}

\item{ARMvar}{column name}

\item{varsup}{: Nature du groupe d’EIs à superposer parmi 3,4 (pour les EIs de grade>= à 3 ou resp 4) ou “Serious” – Note : pas de guillemets pour les chiffres –> par défaut 3}

\item{ARMe}{: indiquer le bras expérimental, il sera placé en premier cad à gauche dans le graphique pour comparer au bras contrôle à droite. Ecrire le label du bras expérimental comme il est dans la base de données (ex: ARMe=“armB”) –> par défaut si cette option n’est pas remplie (donc ARMe=NULL) alors les groupes seront placés dans l’ordre alphabétique des labels}

\item{tri}{: tri selon “pctAll”,“pctVarsup”,“RDAll”,“RDVarsup” toujours en décroissant –> par défaut selon le RD du total “RDAll”}

\item{trivar}{: ex:“armA” label du groupe de traitement selon lequel effectuer le tri –> par défaut NULL tri selon les deux bras}

\item{listcol}{: liste de couleurs pour chacun des groupes de traitements dans un vecteur soit avec les noms implémentés dans R (comme “red”, “blue”…) soit avec des codes (commençant par # par exemple “#52717F”) –> par défaut c(“deepskyblue2”,“red”)}
}
\value{
A plot
}
\description{
Ce graphique est un graphique opposé par le centre donc en forme butterfly comme le graphique empilé précédent, mais celui-ci en est une autre version. Ici nous allons superposée une barre plus fine à la barre principale pour représenter le pourcentage de patients atteints par un EI de grade supérieur ou égale à 3 par rapport au pourcentage de patients atteints par cette EI (tout grade confondu).
La granularité de graphique sera au niveau des SOC, cad qu’une barre représentera le pourcentage de patient ayant eu au moins 1 EI classé dans chaque SOC.
On ajoute à droite de ce graphique un double forest plot pour permettre de quantifier statistiquement les différences de pourcentages entre les groupes présents dans le graphique de gauche.
Note : ici à l’inverse, mais de la même façon que pour le butterfly empilé nous pourrions créer une version pour multiplier ce graphique par SOC et ainsi avoir autant de graphique que de SOC dans l’étude avec cette fois ci cahque barre représentant un EI de la SOC correspondante.
Ce type de graphique n’est adapté que pour deux groupes de traitement.
}
\examples{
library(dplyr)
baseEI <- data.frame(idvar = paste0("Patients", round(runif(n = 100, min = 0, max = 100))),
                     Termsvar = round(runif(n = 100, min = 0, max = 10))) \%>\%
  dplyr::mutate(SOCvar = round(Termsvar/10)) \%>\%
  dplyr::mutate(across(everything(), .fns = as.character)) \%>\%
  mutate(EIdatestart_var = as.Date(runif(n = nrow(.), 1, 200), origin = "2021-01-01"),
         EIdateend_var = as.Date(runif(n = nrow(.), 201, 600), origin = "2021-01-01"),
         gradevar = round(runif(n = nrow(.), 1, 5)))

baseTr  <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  dplyr::mutate(ARMvar = sample(x = c("Placebo", "Treatment"),
                                size = nrow(.),
                                replace = TRUE),
                TTTYN = sample(x = c("Yes", "No"),
                               size = nrow(.),
                               replace = TRUE,
                               prob = c(0.9, 0.1)))

ButterflyBarChartSup(baseEI = baseEI, baseTr = baseTr,
                     idvar = "idvar", SOCvar = "SOCvar", gradevar = "gradevar")

}
