% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButterflyStackAE.R
\name{ButterflyStackAE}
\alias{ButterflyStackAE}
\title{ButterflyStackAE}
\usage{
ButterflyStackAE(
  baseEI,
  baseTr,
  idvar,
  Termsvar,
  SOCvar,
  gradevar,
  TTTYN = NULL,
  ARMvar,
  SOCchoix,
  ARMe = NULL,
  listcol,
  gsup = TRUE,
  grprk = NULL,
  rk = "pctgrade",
  order = "Desc"
)
}
\arguments{
\item{baseEI}{la base des EIs avec :
\itemize{
   \item idvar : l’identifiant du patient : au format character obligatoire
   et sans manquants
   \item Termsvar : le label des PT ou LLT pour chaque EI : au format
   character obligatoire et sans manquants
   \item SOCvar : le label de la SOC associée à chaque EI : au
   format character et sans manquants.
   \item gradevar : le grade de l’EI : au format numeric obligatoire en
   partant de 1 (et allant jusqu’au grade max par exemple 5) et sans
   manquants
}}

\item{baseTr}{la base des traitements avec
\itemize{
   \item idvar : l’identifiant du patient : au format character obligatoire
   et sans manquants
   \item ARMvar : le bras de traitement pour chaque individu : au format
   character obligatoire et sans manquants
   \item TTTYN [optionnal] une variable pour indiquer si le traitement à été pris
   (sert à détecter les patients “à risque”) on recherchera les
   occurences en “Yes” donc doit être codée en “Yes”/“No” en character
}}

\item{idvar}{patient column name}

\item{Termsvar}{PT or LLT AE label column name}

\item{SOCvar}{(character, default = NULL) column name of AE SOC}

\item{gradevar}{AE grade column name}

\item{TTTYN}{Treatment column name}

\item{ARMvar}{Treatment group column name}

\item{SOCchoix}{label de la SOC qui sera représentée (à noter entre
guillemets ex : “BLOOD AND LYMPHATIC SYSTEM DISORDERS”)}

\item{ARMe}{ndiquer le bras expérimental, il sera placé en premier cad à
gauche dans le graphique pour comparer au bras contrôle à droite. Ecrire le
label du bras expérimental comme il est dans la base de données
(ex: ARMe=“armB”)}

\item{listcol}{vecteur de couleurs à indiquer pour chaque grade donc selon le
nombre de modalités de la variable gradevar, doit impérativement être au moins
de la même longeur que le nombre de grades présents dans l’étude (ok si 5
couleurs alors que grades 1 à 4) mais pas moins et contenir les valeurs des
couleurs dans l’ordre des grades (du plus petit au plus grand) –> ex :
c(‘#52717F’, ‘#00A991’, ‘#ffb703’, ‘#CF4B00’, ‘#9a031e’) –> n’a pas de valeur
pas défaut}

\item{gsup}{TRUE/FALSE afficher ou non le graphique résumé dans la partie
supérieure –> par défaut TRUE}

\item{grprk}{“armA”/“armB” variable selon laquelle effectuer le tri, le label
du bras de traitement doit être écrit comme il est présent dans la base –>
par défaut grprk=NULL : tri selon le bras de traitement qui arrive en premier
dans l’ordre alphabétique}

\item{rk}{“pcttot”/“pctgrade” respectivement dans l’ordre décroissant pour la
proportion de patients atteints et dans l’ordre décroissant des grades (cad
place en premier les types d’EIs avec des grades 5 dans l’ordre décroissant
de la proportion de patients avec un grade 5 pour cet EI, puis passe aux
types d’EIs qui n’ont pas de grade 5 mais des grade 4, les tri de la même
façon dans l’ordre décroissant de la proportion de patients avec un grade 4
pour cet EI, et ainsi de suite) –> par défaut “pctgrade”}

\item{order}{“Inc”/“Desc” ordre de d’empilement des grades 1 à 5 ou 5 à 1 à
partir de l’axe central –> par défaut “Desc” car le plus pertinent}
}
\value{
A Butterfly stacked barplot par SOC
}
\description{
Plusieurs versions existent pour ce type de graphique car
“butterfly” caractérise tous les graphiques représentés en miroir/opposition
de deux groupes par le centre du graphique. Dans cette partie on présentera
une comparaison des deux groupes de traitement (en butterfly) avec un
diagramme en barres empilées. Dans cette version, granularité au niveau des
EIs cad on a un butterfly par SOC où chaque barre représente un EI de cette
SOC (voir graphique suivant pour la granularité par SOC cad un graphique
unique avec les SOC pour les barrres).
Beaucoup d’autres variables peuvent être représentées sous cette forme de
graphique, butterfly stacked barplot, mais ici, nous n’allons pas développer
les autres possibilités.
Ce graphique permet de comparer la distribution des EI/EIGs et de leur grades
dans deux groupes de traitement.
Variables : EIs en ligne, proportion de patients en abscisse, variable quali
pour diviser les barres (ici le grade grade ou plutôt le grademax),
opposition par le centre des deux bras de traitement
}
