% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBplot.R
\name{DBplot}
\alias{DBplot}
\title{DBplot}
\usage{
DBplot(
  data1,
  data2,
  data3,
  SOCvar,
  levels_SOC,
  limpct,
  listcol,
  pas,
  nbEvents,
  caption
)
}
\arguments{
\item{data1}{a dataframe}

\item{data2}{another}

\item{data3}{another}

\item{SOCvar}{The SOC variable}

\item{levels_SOC}{The SOC levels}

\item{limpct}{The limpct object}

\item{listcol}{The color list}

\item{pas}{pas}

\item{nbEvents}{nbEvents}

\item{caption}{caption}
}
\value{
A plot
}
\description{
Plot the dumbell
}
