% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DumbbellAE.R
\name{DumbbellAE}
\alias{DumbbellAE}
\title{DumbbellAE}
\usage{
DumbbellAE(
  baseEI,
  baseTr,
  idvar,
  Termsvar,
  SOCvar = NULL,
  TTTYN = NULL,
  ARMvar,
  nbplot = 1,
  nbEvents = FALSE,
  TriSOC = "D",
  TriEI = "D",
  caption = TRUE,
  listcol = c("deepskyblue2", "red")
)
}
\arguments{
\item{baseEI}{la base des EIs avec :
\itemize{
   \item idvar : l’identifiant du patient : au format character obligatoire
   et sans manquants
   \item Termsvar : le label des PT ou LLT pour chaque EI : au format
   character obligatoire et sans manquants
   \item SOCvar [optionnal]: le label de la SOC (SOCvar) associée à chaque
   EI : au format character et sans manquants.
}}

\item{baseTr}{la base des traitements avec
\itemize{
   \item idvar : l’identifiant du patient : au format character obligatoire
   et sans manquants
   \item ARMvar : le bras de traitement pour chaque individu : au format
   character obligatoire et sans manquants
   \item TTTYN [optionnal] une variable pour indiquer si le traitement à été pris
   (sert à détecter les patients “à risque”) on recherchera les
   occurences en “Yes” donc doit être codée en “Yes”/“No” en character
}}

\item{idvar}{(character) column name of patient id column}

\item{Termsvar}{(character) column name of PT or LLT label for each AE}

\item{SOCvar}{(character, default = NULL) column name of AE SOC}

\item{TTTYN}{(character, default = NULL) column name indicating if the
treatemnt was taken}

\item{ARMvar}{(character) column name for treatment group}

\item{nbplot}{(numeric, default = 1) nombre de graphiques à créer
(l’utilisateur doit choisir en fonction du nombre d’EI différents dans la
base de données) cad division du graphique global en plusieurs sous
graphiques lorsque le nombre d’EI est top grand.}

\item{nbEvents}{(boolean, default = FALSE) Choisir d’afficher le nombre
d’évènement brute par PT (ou LLT) sous forme de deux colonnes de chiffres
supplémentaires à la droite du graphique}

\item{TriSOC}{(character, default = "D) “C”/“D”/“alpha” pour faire un tri
croissant, décroissant ou selon le RD ou alphabétique des SOC}

\item{TriEI}{(character, default = "D) “C”/“D”/“alpha” idem pour PT (ou LLT)}

\item{caption}{(boolean, default = TRUE) TRUE/FALSE afficher une note de base
de graphique avec des informations supplémentaires sur le format et le
continu du graphique}

\item{listcol}{(vector of character, default = (“red”,“deepskyblue2”)) liste
de couleurs pour chacun des groupes de traitements dans un vecteur soit avec
les noms implémentés dans R (comme “red”, “blue”…) soit avec des codes
(commençant par # par exemple “#52717F”)}
}
\value{
A Dumbell + forest plots
}
\description{
Représente pour chaque EI :
\itemize{
   \item dans la partie gauche (Dumbbell plot) : représente soit le nombre
   brut de patients ayant eu au moins 1 fois l’EI soit la proportion de
   patients ayant eu cet EI - surtout lorsque les groupes ne sont pas
   égaux - avec 2 points par ligne/par EI (1 pour chaque groupe de traitement
   différenciés par des formes et couleurs). Note : Ne compte qu’une
   occurrence de chaque EI par patient, d’où l’utilité éventuelle d’ajouter
   deux colonnes au graphique avec le nombre total d’évènements par type EI
   et par groupe (en faisant toutefois attention à la taille des deux groupes
   de traitement pour comparer ces chiffres)
   \item dans la partie droite (Forest plot) : le Risk Difference (RD)
   calculé pour chaque type d’EI avec l’IC à 95%. Dans cette partie du
   graphique, il est aussi possible d’utiliser, sous la même forme, le risque
   relatif (RR) ou encore l’odd ratio (OR). Pour ce dernier on choisira une
   échelle logarithmique. –> Plus visuel que d’avoir seulement les chiffres
   et IC dans un tableau, ici on voit directement les IC qui se chevauchent
   ou non, ou ceux qui contiennent la valeur 0 (pour le RD) par exemple.
}
L’ajout du Forest plot à un Dumbbell plot classic est essentiel car il permet
d’interpréter au mieux l’écart entre les deux groupes et de quantifier
certains aspects du jeu de données. L’addition de ces deux graphiques permet
d’identifier facilement les différences significatives entre les bras de
traitement.
Ce graphique est adapté à seulement deux groupes de traitement.
}
