% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LasagnaAE.R
\name{LasagnaAE}
\alias{LasagnaAE}
\title{LasagnaAE}
\usage{
LasagnaAE(data, unit, choixEI, vect_grade, listcol, idpat)
}
\arguments{
\item{data}{dataframe}

\item{unit}{unit}

\item{choixEI}{choixEI}

\item{vect_grade}{vect_grade}

\item{listcol}{listcol}

\item{idpat}{idpat}
}
\value{
A plot
}
\description{
helper for PanelLasagnaAlluvialAE
}
