% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineErrorBars.R
\name{LineErrorBars}
\alias{LineErrorBars}
\title{LineErrorBars}
\usage{
LineErrorBars(
  baseEI,
  baseTr,
  baseDates,
  idvar,
  Termsvar,
  gradevar,
  EIdatestart_var,
  EIdateend_var,
  TTTYN = NULL,
  ARMvar,
  tttdebdate_var,
  tttfindate_var,
  visitedate_var = NULL,
  visitnum_var = NULL,
  grd = 2,
  choixEI,
  unit,
  suivi = TRUE,
  listcol = c("red", "deepskyblue2"),
  caption = TRUE
)
}
\arguments{
\item{baseEI}{: la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item la date de début de l’EI (EIdatestart_var) : au format Date (important) obligatoire
   \item la date de fin de l’EI (EIdateend_var) : au format Date (important) obligatoire
   \item le grade de l’EI (gradevar) : au format numeric obligatoire en partant de 1 (et allant jusqu’au grade max par exemple 5) et sans manquants
}}

\item{baseTr}{: la base des traitements avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item [non obligatoire] une variable pour indiquer si le traitement à été pris (sert à détecter les patients “à risque”) TTTYN on recherchera les occurences en “Yes” donc doit être codée en “Yes”/“No” en character
   \item [non obligatoire] une variable pour le numéro de la visite (ou du cycle) (visitenum_var): non obligatoire, à indiquer seulement s’il y a un format de visites ou de cycles dans l’étude. Au format numeric et sans manquants.
   \item [non obligatoire] combinée à, la date de début de ce cycle (date de visite, de prise de traitement en début de période par exemple) (visitedate_var) : idem non obligatoire, à indiquer seulement s’il y a format de visites ou de cycles dans l’étude. Au format Date et peut avoir des manquantes si par exmple il n’y a pas eu de prise de traitement à ce cycle là (on aurait TTTYN=“No”).
   \item Note 2: il est possible d’avoir plusieurs dates par cycles/période si par exemple on a plusieurs injections ou plusieurs prises de traitement entre deux visites ou dans un même cycle. Pour choisir la date de début de cycle, nous prendrons la plus ancienne.
}}

\item{baseDates}{: la base qui contient les dates principales liées à chaque patient, celles qui nous sont utiles pour ce graphique sont :
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item la date de début de traitement tttdebdate_var : au format Date obligatoire et sans manquants
   \item la date de fin de traitement/fin de suivi tttfindate_var qui correspond à la fin de la période traitement par exemple, des EIs peuvent se situer après cette date. Si le format “cycle” est choisit on arrête le dernier cycle à cette date, si un autre format est choisit, la date de fin du graphique sera en fonction de l’option suivi. Si suivi=TRUE alors on prend tous les EIs enregistrés, sinon si suivi=False alors on termine l’axe du temps à la date de fin tttfindate_var. Au format Date et sans manqants (si on choisit l’option suivi=FALSE ou unit="cycle")
}}

\item{idvar}{column name}

\item{Termsvar}{column name}

\item{gradevar}{column name}

\item{EIdatestart_var}{column name}

\item{EIdateend_var}{column name}

\item{TTTYN}{column name}

\item{ARMvar}{column name}

\item{tttdebdate_var}{column name}

\item{tttfindate_var}{column name}

\item{visitedate_var}{column name}

\item{visitnum_var}{column name}

\item{grd}{: comparaison entre les deux groupe de la proportion de patients ayant eu au moins un EI de grade >= grd par rapport aux pct de patient ayant eu l’EI (tout grade confondu) (d’où ne peut pas être égal à 1) –> par défaut 2}

\item{choixEI}{: EI choisis à afficher sur le graphique –> ex “ANAEMIA”}

\item{unit}{: unité au choix entre “cycle”/“week”/“month”/“quarter”/“halfyear”/“year”. Comme précisé précédement le cycle peut correspondre à une période entre deux visites ou deux prise de traitement en fonction de l’étude… –> pas de valeur par défaut car dépend de chaque étude.}

\item{suivi}{: TRUE / FALSE afficher ou non les EIs survenus après la date de fin de traitement / début de suivi (tttfindate_var) –> par défaut TRUE}

\item{listcol}{: liste de couleurs pour chacun des groupes de traitements dans un vecteur soit avec les noms implémentés dans R (comme “red”, “blue”…) soit avec des codes (commençant par # par exemple “#52717F”) –> par défaut c(“red”,“deepskyblue2”)}

\item{caption}{: TRUE/FALSE afficher une note de base de graphique avec des informations supplémentaires sur le format et le continu du graphique –> par défaut TRUEtion}
}
\value{
A plot
}
\description{
Ce graphique peut être combiné avec le panel Lasagna/Alluvial pour comparer avec un IC les pourcentage des patients atteints par des EIs de grade >= value.
Il est composé d’une barre d’erreur avec le RD et son IC à 95% pour chaque période (choix entre cycle, semaine, mois…) et permet par exemple de voir si la proportion de patient ayant eu une Anémie de grade >= 3 (parmis ceux ayant eu une Anémie) est significativement plus élevée dans un groupe de traitement ou non.
Ce graphique n’est adapté que pour deux groupes de traitements.
}
\examples{
library(dplyr)
baseEI <- data.frame(idvar = paste0("Patients", round(runif(n = 100, min = 0, max = 100))),
                     Termsvar = round(runif(n = 100, min = 0, max = 2))) \%>\%
  dplyr::mutate(SOCvar = round(Termsvar/10)) \%>\%
  dplyr::mutate(across(everything(), .fns = as.character)) \%>\%
  mutate(EIdatestart_var = as.Date(runif(n = nrow(.), 1, 200), origin = "2021-01-01"),
         EIdateend_var = as.Date(runif(n = nrow(.), 201, 600), origin = "2021-01-01"),
         gradevar = round(runif(n = nrow(.), 1, 5)))

baseTr  <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  dplyr::mutate(ARMvar = sample(x = c("Placebo", "Treatment"),
                                size = nrow(.),
                                replace = TRUE),
                TTTYN = sample(x = c("Yes", "No"),
                               size = nrow(.),
                               replace = TRUE,
                               prob = c(0.9, 0.1)))

baseDates <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  mutate(tttdebdate_var = as.Date(runif(n = nrow(.), -100, 0), origin = "2021-01-01"),
         tttfindate_var = as.Date(runif(n = nrow(.), 201, 600), origin = "2021-01-01"))

LineErrorBars(
  baseEI = baseEI,
  baseTr = baseTr,
  baseDates = baseDates,
  idvar = "idvar",
  Termsvar = "Termsvar",
  EIdatestart_var = "EIdatestart_var",
  EIdateend_var = "EIdateend_var",
  ARMvar = "ARMvar",
  tttdebdate_var = "tttdebdate_var",
  tttfindate_var = "tttfindate_var",
  choixEI = 1,
  unit = "month",
  listcol = viridis::viridis(n = 5)
)

}
