% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBioAE_function_Vfin.R
\name{PanelBioAE}
\alias{PanelBioAE}
\title{PanelBioAE}
\usage{
PanelBioAE(
  baseBio,
  idvar,
  biodate_var,
  unitbio_var,
  biolib_var,
  labvisite_var = NULL,
  biores_var,
  USU,
  list_Bio = NULL,
  suivi = TRUE
)
}
\arguments{
\item{baseBio}{la base des biologies avec les résultats des différentes biologies pour chaque patient
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item la date du relevé biodate_var : au format Date obligatoire et sans manquants
   \item le label de la biologie réalisée biolib_var : au format character obligatoire et sans manquants peuvent être tous identique si une seule biologie dans l’étude
   \item l’unité pour chaque type de biologie unitbi_var : au format character obligatoire et sans manquants
   \item le résultats de biologie biores_var : en numéric obligatoire
   \item [non obligatoire] le label de la visite pendant laquelle a été faite la bio labvisite_var : au format character sans manquants
}}

\item{idvar}{column name}

\item{biodate_var}{column name}

\item{unitbio_var}{column name}

\item{biolib_var}{column name}

\item{labvisite_var}{column name}

\item{biores_var}{column name}

\item{USU}{: choix du patient pour lequel on souhaite voir les résultats de biologie –> comme présent dans la base ex : “1002”}

\item{list_Bio}{: liste avec les labels de biologies à afficher - ex : c(“ALAT”,“ASAT”) –> par défaut NULL ,pas de bio indiqué alors on les affiche tous}

\item{suivi}{: TRUE/FALSE afficher les résultats de bio mesurée après la fin du traitement/début du suivi –> par défaut TRUE}
}
\value{
A plot
}
\description{
Ce graphique représente de façon classique les résultats de biologies d’un patient au cours du temps (plusieurs unités possibles), il se combine bien avec le Patient Span Chart qui représente les EIs au cours du temps pour mettre en regards les biologies relevées et les évènements indésirables.
Pour ce type de graphique il peut y avoir n’importe quel nombre de bras de traitement car on ne représente qu’un patient à la fois pour lequel on indique le bras dans lequel il est.
}
