% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientSpanAE.R
\name{PatientSpanAE}
\alias{PatientSpanAE}
\title{PatientSpanAE}
\usage{
PatientSpanAE(
  baseEI,
  baseTr,
  baseDates,
  baseBio,
  idvar,
  Termsvar,
  SOCvar,
  EIdatestart_var,
  EIdateend_var,
  gradevar,
  ongovar,
  outvar,
  SAEvar,
  ttt1var,
  ttt2var,
  ARMvar,
  tttlab_var,
  tttdate_var,
  visnum_var,
  tttdebdate_var,
  tttfindate_var,
  randodate_var,
  screendate_var,
  datebiovar,
  labbiovar,
  resbiovar,
  unitbiovar,
  choixUSU,
  suivi = TRUE,
  listcol,
  bio = TRUE
)
}
\arguments{
\item{baseEI}{: la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item le label de la SOC (SOCvar) associée à chaque EI : au format character et sans manquants.
   \item la date de début de l’EI (EIdatestart_var) : au format Date (important) obligatoire
   \item la date de fin de l’EI (EIdateend_var) : au format Date (important) obligatoire
   \item le grade de l’EI (gradevar) : au format numeric obligatoire en partant de 1 (et allant jusqu’au grade max par exemple 5) et sans manquants
   \item une variable pour inditifier si l’EI est toujours en cours (ongovar) : “Yes”/“No” au format character ou facteur obligatoire
   \item une variable par traitement pour indiquer si l’EI est lié au traitement (ttt1var1,ttt2var): ici codé en “Yes”/“No” pour chaque molécule obligatoire
   \item si l’EI est de type sérieux (SAEvar) : codé en “Yes”/“No” obligatoire
   \item une variable qui indique l’outcome de l’EI (outvar) : codée avec les différents outcome possibles soit en character soit en facteur obligatoire
}
Note : si la date de fin n’est pas renseignée pour certains EIs alors ceux-ci ne seront pas affichés dans le graphique}

\item{baseTr}{la base des traitements avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item le label pour chacun des traitements administrés tttlab (car possible plusieurs molécules/traitements dans chaque bras) : en format character ou facteur
}}

\item{baseDates}{la base qui contient les dates principales liées à chaque patient, celles qui nous sont utiles pour ce graphique sont :
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item la date de début de traitement tttdebdate_var : au format Date obligatoire et sans manquants
   \item la date de fin de traitement/fin de suivi tttfindate_var qui correspond à la fin de la période traitement par exemple, des EIs peuvent se situer après cette date. Si le format “cycle” est choisit on arrête le dernier cycle à cette date, si un autre format est choisit, la date de fin du graphique sera en fonction de l’option suivi. Si suivi=TRUE alors on prend tous les EIs enregistrés, sinon si suivi=False alors on termine l’axe du temps à la date de fin tttfindate_var. Au format Date et sans manqants (si on choisit l’option suivi=FALSE ou unit="cycle")
}}

\item{baseBio}{base qui contient les résultats de biologies des patients dans laquelle il faut le numéro du patient idvar, la date de la biologie (datebiovar) en format Date, le label de la biologie (labbiovar) ici on prendra l’exemple de l’hémoglobine, le résultat (resbiovar) et enfin l’unité de la mesure de biologie (unitbiovar).}

\item{idvar}{column name}

\item{Termsvar}{column name}

\item{SOCvar}{column name}

\item{EIdatestart_var}{column name}

\item{EIdateend_var}{column name}

\item{gradevar}{column name}

\item{ongovar}{column name}

\item{outvar}{column name}

\item{SAEvar}{column name}

\item{ttt1var}{column name}

\item{ttt2var}{column name}

\item{ARMvar}{column name}

\item{tttlab_var}{column name}

\item{tttdate_var}{column name}

\item{visnum_var}{column name}

\item{tttdebdate_var}{column name}

\item{tttfindate_var}{column name}

\item{randodate_var}{column name}

\item{screendate_var}{column name}

\item{datebiovar}{column name}

\item{labbiovar}{column name}

\item{resbiovar}{column name}

\item{unitbiovar}{column name}

\item{choixUSU}{: numéro du patient (à noter entre guillemets ex : “1” carla variable idvar doit être au format character) comme il est présent dans la base}

\item{suivi}{: TRUE/FALSE afficher ou non les EIs ayant eu lieu après la fin du traitement/fin de suivi (tttfindate_var) (automatiquement FALSE si on choisit l’unité “cycle”) –> par défaut FALSE}

\item{listcol}{: vecteur de couleurs à indiquer pour chaque grade donc selon le nombre de modalités de la variable gradevar, doit impérativement être au moins de la même longeur que le nombre de grades présents dans l’étude (ok si 5 couleurs alors que grades 1 à 4) mais pas moins et contenir les valeurs des couleurs dans l’ordre des grades (du plus petit au plus grand) –> ex : c(‘#52717F’, ‘#00A991’, ‘#ffb703’, ‘#CF4B00’, ‘#9a031e’) –> n’a pas de valeur pas défaut}

\item{bio}{: TRUE/FALSE afficher les résultats de bio (dans l’ex hémoglobine) ou non –> par défaut TRUE}
}
\value{
A plot
}
\description{
Permet d’avoir sur un même graphique pour un patient données les informations sur sa prise de traitement et sur ses EIs afin de permettre une lecture globale du profil de toxicité de chaque patient.
}
