% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_las.R
\name{ranking_las}
\alias{ranking_las}
\title{ranking_las}
\usage{
ranking_las(table_las)
}
\arguments{
\item{table_las}{A table to rank}
}
\value{
A ranked table
}
\description{
Ranking de chacune des tables pour les graphiques used by
PanelLasagnaAlluvialAE()
}
