% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunct.R
\name{RDfunct}
\alias{RDfunct}
\title{RDfunct}
\usage{
RDfunct(a, b, N1, N0, CRC = FALSE, conf.level = 0.95)
}
\arguments{
\item{a}{The number of disease occurence among exposed cohort.}

\item{b}{The number of disease occurence among non-exposed cohort.}

\item{N1}{The population at risk of the exposed cohort.}

\item{N0}{The population at risk of the unexposed cohort.}

\item{CRC}{Logical. If TRUE, calculate confidence intervals for each risk. Default is FALSE.}

\item{conf.level}{Probability for confidence intervals. Default is 0.95.}
}
\value{
A list with estimate, conf.int and p.value
}
\description{
calcul du RD (pris de fmsb::riskdifference)
}
