% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TendrilAE.R
\name{TendrilAE}
\alias{TendrilAE}
\title{TendrilAE}
\usage{
TendrilAE(
  baseEI,
  baseTr,
  baseDates,
  idvar,
  Termsvar,
  EIdatestart_var,
  ARMvar,
  tttdebdate_var,
  coltype = NULL,
  caption = TRUE
)
}
\arguments{
\item{baseEI}{la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item la SOC correspondante (SOC_COD) : label (et non code) au format character obligatoire et sans manquants
   \item la date de début de l’EI (EIdatestart_var) : au format Date (important) obligatoire
}}

\item{baseTr}{la base des traitements avec*
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item [non obligatoire] une variable pour indiquer si le traitement à été pris (sert à déte
}}

\item{baseDates}{la base qui contient les dates principales liées à chaque patient, celles qui nous sont utiles pour ce graphique sont :
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item la date de début de traitement tttdebdate_var : au format Date obligatoire et sans manquants
   \item la date de fin de traitement/fin de suivi tttfindate_var qui correspond à la fin de la période traitement par exemple, des EIs peuvent se situer après cette date. Si le format “cycle” est choisit on arrête le dernier cycle à cette date, si un autre format est choisit, la date de fin du graphique sera en fonction de l’option suivi. Si suivi=TRUE alors on prend tous les EIs enregistrés, sinon si suivi=False alors on termine l’axe du temps à la date de fin tttfindate_var. Au format Date et sans manqants (si on choisit l’option suivi=FALSE ou unit="cycle")
}}

\item{idvar}{column name}

\item{Termsvar}{column name}

\item{EIdatestart_var}{column name}

\item{ARMvar}{column name}

\item{tttdebdate_var}{column name}

\item{coltype}{: “p.adj”,“fish”,“p”,“rdiff”,“OR”,“RR”,“FDR.tot”,“TermsCount” type de colorisation (options déjà présentes pour les fonction liées au package du Tendril) –> par défaut NULL qui donne une couleur à chaque branche color selon chaque Term}

\item{caption}{: TRUE/FALSE afficher une note de base de graphique avec des informations supplémentaires sur le format et le continu du graphique –> par défaut TRUE}
}
\value{
A plot
}
\description{
Permet de montrer à la fois l’importance relative des risques et le pattern temporel des EIs durant l’étude. Visualisation qui résume le profil global des EIs d’une étude, montre les EIs d’importance majeure et illustre l’évolution temporelle des EIs reportés. Montre TOUS les EIs, même les récurrents (occurrence multiple pour chaque patient), et pas seulement la première occurrence comme c’est le cas pour d’autres graphiques.
}
\examples{
library(dplyr)
baseEI <- data.frame(idvar = paste0("Patients", round(runif(n = 100, min = 0, max = 100))),
                     Termsvar = round(runif(n = 100, min = 0, max = 2))) \%>\%
  dplyr::mutate(SOCvar = round(Termsvar/10)) \%>\%
  dplyr::mutate(across(everything(), .fns = as.character)) \%>\%
  mutate(EIdatestart_var = as.Date(runif(n = nrow(.), 1, 200), origin = "2021-01-01"),
         EIdateend_var = as.Date(runif(n = nrow(.), 201, 600), origin = "2021-01-01"),
         gradevar = round(runif(n = nrow(.), 1, 5)))

baseTr  <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  dplyr::mutate(ARMvar = sample(x = c("Placebo", "Treatment"),
                                size = nrow(.),
                                replace = TRUE),
                TTTYN = sample(x = c("Yes", "No"),
                               size = nrow(.),
                               replace = TRUE,
                               prob = c(0.9, 0.1)))

baseDates <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  mutate(tttdebdate_var = as.Date(runif(n = nrow(.), -100, 0), origin = "2021-01-01"),
         tttfindate_var = as.Date(runif(n = nrow(.), 201, 600), origin = "2021-01-01"))

TendrilAE(baseEI = baseEI, baseTr = baseTr, baseDates = baseDates,
          idvar = "idvar", Termsvar = "Termsvar", EIdatestart_var = "EIdatestart_var",
          ARMvar = "ARMvar", tttdebdate_var = "tttdebdate_var")

}
