% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TendrilAETimeSeries.R
\name{TendrilAETimeSeries}
\alias{TendrilAETimeSeries}
\title{TendrilAETimeSeries}
\usage{
TendrilAETimeSeries(
  baseEI,
  baseTr,
  baseDates,
  idvar,
  Termsvar,
  EIdatestart_var,
  ARMvar,
  tttdebdate_var,
  CODlist = NULL,
  listcol = NULL
)
}
\arguments{
\item{baseEI}{la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item la SOC correspondante (SOC_COD) : label (et non code) au format character obligatoire et sans manquants
   \item la date de début de l’EI (EIdatestart_var) : au format Date (important) obligatoire
}}

\item{baseTr}{la base des traitements avec*
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item [non obligatoire] une variable pour indiquer si le traitement à été pris (sert à déte
}}

\item{baseDates}{la base qui contient les dates principales liées à chaque patient, celles qui nous sont utiles pour ce graphique sont :
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item la date de début de traitement tttdebdate_var : au format Date obligatoire et sans manquants
   \item la date de fin de traitement/fin de suivi tttfindate_var qui correspond à la fin de la période traitement par exemple, des EIs peuvent se situer après cette date. Si le format “cycle” est choisit on arrête le dernier cycle à cette date, si un autre format est choisit, la date de fin du graphique sera en fonction de l’option suivi. Si suivi=TRUE alors on prend tous les EIs enregistrés, sinon si suivi=False alors on termine l’axe du temps à la date de fin tttfindate_var. Au format Date et sans manqants (si on choisit l’option suivi=FALSE ou unit="cycle")
}}

\item{idvar}{column name}

\item{Termsvar}{column name}

\item{EIdatestart_var}{column name}

\item{ARMvar}{column name}

\item{tttdebdate_var}{column name}

\item{CODlist}{: liste de PT à mettre en valeur dans le graph (les autres seront tracés en plus clair en arrière plan si format=“All” ou non présentes si format=“list”), ex c(“ANAEMIA”,“NAUSEA”,“ASTHENIA”) –> par défaut NULL tous les EIs de la base de données sont affichés}

\item{listcol}{: vecteur/liste de couleurs à indiquer pour colorer la selection de Terms (dans CODlist), ex (lié à l’exemple pour CODlist) : c(“purple”,“orange”,“green”) –> par défaut NULL, si CODlist NON null alors tous les EIs selectionnés dans COD seront en noir (par rapport aux autres non selectionnés en gris)}
}
\value{
A plot
}
\description{
En prenant donc les mêmes informations que pour le Tendril plot, c’est a dire les EIs, le bras de traitement et le jours d’apparition de chaque EI on construit un Time Series plot avec en ordonné la différence d’occurence entre chaque groupe, et en abcisse le nombre de jours écoulés depuis le début du traitement.
}
