% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeToSpanChart.R
\name{TimeToSpanChart}
\alias{TimeToSpanChart}
\title{TimeToSpanChart}
\usage{
TimeToSpanChart(
  baseEI,
  baseTr,
  baseDates,
  idvar,
  Termsvar,
  EIdatestart_var,
  EIdateend_var,
  gradevar,
  TTTYN = NULL,
  ARMvar,
  tttdebdate_var,
  tttfindate_var,
  listEI = NULL,
  unit = "weeks",
  caption = TRUE,
  listcol = c("deepskyblue2", "red")
)
}
\arguments{
\item{baseEI}{la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item le grade de l’EI (gradevar) : au format numeric obligatoire en partant de 1 (et allant jusqu’au grade max par exemple 5) et sans manquants
   \item la date de début de l’EI (EIdatestart_var) : au format Date (important) obligatoire
   \item la date de fin de l’EI (EIdateend_var) : au format Date (important) obligatoire
}}

\item{baseTr}{la base des traitements avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item [non obligatoire] une variable pour indiquer si le traitement à été pris (sert à détecter les patients “à risque”) TTTYN on recherchera les occurences en “Yes” donc doit être codée en “Yes”/“No” en character
}}

\item{baseDates}{la base qui contient les dates principales liées à chaque patient, celles qui nous sont utiles pour ce graphique sont :
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item la date de début de traitement tttdebdate_var : au format Date obligatoire et sans manquants
   \item la date de fin de traitement/fin de suivi tttfindate_var qui correspond à la fin de la période traitement par exemple, des EIs peuvent se situer après cette date. Si le format “cycle” est choisit on arrête le dernier cycle à cette date, si un autre format est choisit, la date de fin du graphique sera en fonction de l’option suivi. Si suivi=TRUE alors on prend tous les EIs enregistrés, sinon si suivi=False alors on termine l’axe du temps à la date de fin tttfindate_var. Au format Date et sans manqants (si on choisit l’option suivi=FALSE ou unit="cycle")
}}

\item{idvar}{column name}

\item{Termsvar}{column name}

\item{EIdatestart_var}{column name}

\item{EIdateend_var}{column name}

\item{gradevar}{column name}

\item{TTTYN}{column name}

\item{ARMvar}{column name}

\item{tttdebdate_var}{column name}

\item{tttfindate_var}{column name}

\item{listEI}{: liste des EIs choisis à afficher sur le graphique –> ex c(“ANAEMIA”,“ASTHENIA”,“NAUSEA”) –> par défaut NULL prendra les 20 premiers en nombre total d’évènements}

\item{unit}{: unité au choix entre “days”/“weeks”/“month” –> par défaut “weeks”}

\item{caption}{: TRUE/FALSE afficher une note de base de graphique avec des informations supplémentaires sur le format et le continu du graphique –> par défaut TRUE}

\item{listcol}{: liste de couleurs pour chacun des groupes de traitements dans un vecteur soit avec les noms implémentés dans R (comme “red”, “blue”…) soit avec des codes (commençant par # par exemple “#52717F”) –> par défaut c(“deepskyblue2”,“red”)}
}
\value{
A plot
}
\description{
Ce graphique permet d’observer et surtout de comparer la temporalité de plusieurs aspects d’un groupe d’EI (une vingtaine) par rapport à la durée du traitement dans chaque bras. On y observe le temps médian d’appartition du premier EI sa durée et le temps médian d’appartition d’un grade supérieur à 3 et d’un grade supérieur à 4.
}
