% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreemapAE.R
\name{TreemapAE}
\alias{TreemapAE}
\title{TreemapAE}
\usage{
TreemapAE(
  baseEI,
  baseTr,
  idvar,
  Termsvar,
  SOCvar = NULL,
  gradevar,
  TTTYN = NULL,
  ARMvar,
  choixSOC = NULL,
  listcol,
  ARMe = NULL,
  caption = TRUE
)
}
\arguments{
\item{baseEI}{la base des EIs avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le label des PT ou LLT (Termsvar) pour chaque EI : au format character obligatoire et sans manquants
   \item la SOC correspondante (SOC_COD) : label (et non code) au format character obligatoire et sans manquants
   \item le grade de l’EI (gradevar) : au format numeric obligatoire en partant de 1 (et allant jusqu’au grade max par exemple 5) et sans manquants
}}

\item{baseTr}{la base des traitements avec
\itemize{
   \item l’identifiant du patient (idvar) : au format character obligatoire et sans manquants
   \item le bras de traitement pour chaque individu (ARMvar) : au format character obligatoire et sans manquants
   \item [non obligatoire] une variable pour indiquer si le traitement à été pris (sert à détecter les patients “à risque”) TTTYN on recherchera les occurences en “Yes” donc doit être codée en “Yes”/“No” en character
}}

\item{idvar}{column name}

\item{Termsvar}{column name}

\item{SOCvar}{column name}

\item{gradevar}{column name}

\item{TTTYN}{column name}

\item{ARMvar}{column name}

\item{choixSOC}{: permet de choisir de zoomer sur une SOC en particulier (ex : “BLOOD AND LYMPTAHIC SYSTEM DISORDERS”) –> par défaut NULL affiche toutes les SOC}

\item{listcol}{: vecteur de couleurs à indiquer pour chaque grade donc selon le nombre de modalités de la variable gradevar, doit impérativement être au moins de la même longeur que le nombre de grades présents dans l’étude (ok si 5 couleurs alors que grades 1 à 4) mais pas moins et contenir les valeurs des couleurs dans l’ordre des grades (du plus petit au plus grand) –> ex : c(‘#52717F’, ‘#00A991’, ‘#ffb703’, ‘#CF4B00’, ‘#9a031e’) –> n’a pas de valeur pas défaut}

\item{ARMe}{: indiquer le bras expérimental, il sera placé en premier cad à gauche dans le graphique pour comparer au bras contrôle à droite. Ecrire le label du bras expérimental comme il est dans la base de données (ex: ARMe=“armB”) –> par défaut si cette option n’est pas remplie (donc ARMe=NULL) alors les groupes seront placés dans l’ordre alphabétique des labels“}

\item{caption}{: TRUE/FALSE afficher une note de base de graphique avec des informations supplémentaires sur le format et le continu du graphique –> par défaut TRUE}
}
\value{
A plot
}
\description{
Pour ce graphique nous avons un rectangle par groupe de traitement divisé en plusieurs sous rectangles et colorés selon une échelle de couleur, le but est de comparer les proportions de chacun des sous-rectangles entre les groupes et leur couleurS.
Ici on divise chaque rectangle en sous rectangle selon les SOC puis selon chaque type d’EIs et on compare la taille par exemple du groupe de rectangle pour une SOC donnée entre les bras de traitement ou la taille d’un rectangle pour un EI entre les groupes de traitement ou alors la couleur globale des rectangles…
Les rectangles sont automatiquement ordonnés du coin inférieur gauche au coin supérieur droit dans l’ordre croissant selon leur taille.
Note : ce sont des comparaisons globales et non des comparaison précises avec des axes, il faudra creuser les hypothèses sorties de ce graphiques avec d’autres graphiques de ce documents ou en explirant les données
Ce graphique permet de comparer deux groupes de traitements car chacun à un rectangle qui lui est propre. Dans le cas où l’étude comporte plus de deux groupes de traitements, il est possible de faire un premier graphique avec les deux premiers bras puis de réitérer l’opération avec les autres bras de traitements.
}
