% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoAE.R
\name{VolcanoAE}
\alias{VolcanoAE}
\title{VolcanoAE}
\usage{
VolcanoAE(
  baseEI,
  baseTr,
  idvar,
  Termsvar,
  TTTYN = NULL,
  ARMvar,
  caption = TRUE,
  listcol = c("red", "deepskyblue2")
)
}
\arguments{
\item{baseEI}{The base of AEs with:
\itemize{
   \item idvar: Patient identifier, in character format, mandatory and without missing values.
   \item Termsvar: PT or LLT label for each AE, in character format, mandatory and without missing values.
}}

\item{baseTr}{The base of treatments with:
\itemize{
   \item idvar: Patient identifier, in character format, mandatory and without missing values.
   \item ARMvar: Treatment arm for each individual, in character format, mandatory and without missing values.
   \item TTTYN : A variable indicating if the treatment was taken (used to detect "at-risk" patients). Occurrences will be searched for "Yes", so it should be coded as "Yes"/"No" in character format.
}}

\item{idvar}{(character) Column name of the patient identifier.}

\item{Termsvar}{(character) Column name of the label for each AE.}

\item{TTTYN}{(character, default = NULL) Column name indicating if the treatment was taken.}

\item{ARMvar}{(character) Column name for the treatment group.}

\item{caption}{(boolean, default = TRUE) TRUE/FALSE to display a basic graphic note with additional information about the format and content of the graph.}

\item{listcol}{(vector of character, default = (“red”,“deepskyblue2”)) A list of colors for each treatment group, provided as a vector with either R's implemented color names (e.g., “red”, “blue”...) or color codes (starting with #, e.g., “#52717F”).}
}
\value{
A Volcano plot
}
\description{
Each bubble represents an Adverse Event (AE), with the bubble's size representing the occurrence of the AE (either the number of AEs or the number of patients who experienced the AE). The color of the bubbles corresponds to the treatment group. Fisher's p-value is shown on the y-axis and determines the saturation of the bubbles. On the x-axis, a risk difference between the two groups is displayed.
}
\details{
VolcanoAE
}
\examples{
library(dplyr)
baseEI <- data.frame(idvar = paste0("Patients", round(runif(n = 100, min = 0, max = 100))),
                     Termsvar = round(runif(n = 100, min = 0, max = 20))) \%>\%
  dplyr::mutate(SOCvar = round(Termsvar/10)) \%>\%
  dplyr::mutate(across(everything(), .fns = as.character))

baseTr  <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  dplyr::mutate(ARMvar = sample(x = c("Placebo", "Treatment"),
                                size = nrow(.),
                                replace = TRUE),
                ARMvar = factor(ARMvar, levels = c("Treatment", "Placebo")),
                TTTYN = sample(x = c("Yes", "No"),
                               size = nrow(.),
                               replace = TRUE,
                               prob = c(0.9, 0.1)))

VolcanoAE(baseEI = baseEI,
          baseTr = baseTr,
          idvar = "idvar",
          Termsvar = "Termsvar",
          TTTYN = "TTTYN",
          ARMvar = "ARMvar")

}
