% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarChart.R
\name{BarChart}
\alias{BarChart}
\title{BarChart}
\usage{
BarChart(data, group, Type = BPType)
}
\arguments{
\item{data}{A dataframe}

\item{group}{A groupe}

\item{Type}{A type}
}
\value{
A plot
}
\description{
helper for PanelLasagnaAlluvialAE
}
