% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBplot.R
\name{DBplot}
\alias{DBplot}
\title{DBplot}
\usage{
DBplot(data1, data2, data3)
}
\arguments{
\item{data1}{a dataframe}

\item{data2}{another}

\item{data3}{another}
}
\value{
A plot
}
\description{
Plot the dumbell
}
