% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoAE.R
\name{VolcanoAE}
\alias{VolcanoAE}
\title{VolcanoAE}
\usage{
VolcanoAE(
  baseEI,
  baseTr,
  idvar,
  Termsvar,
  TTTYN = NULL,
  ARMvar,
  seuillab = 0.01,
  caption = TRUE,
  listcol = c("red", "deepskyblue2")
)
}
\arguments{
\item{baseEI}{la base des EIs avec :
\itemize{
   \item idvar : l’identifiant du patient : au format character obligatoire
   et sans manquants
   \item Termsvar : le label des PT ou LLT pour chaque EI : au format
   character obligatoire et sans manquants
}}

\item{baseTr}{la base des traitements avec
\itemize{
   \item idvar : l’identifiant du patient : au format character obligatoire
   et sans manquants
   \item ARMvar : le bras de traitement pour chaque individu : au format
   character obligatoire et sans manquants
   \item TTTYN [optionnal] une variable pour indiquer si le traitement à été pris
   (sert à détecter les patients “à risque”) on recherchera les
   occurences en “Yes” donc doit être codée en “Yes”/“No” en character
}}

\item{idvar}{(character) column name of patient id column}

\item{Termsvar}{(character) column name of PT or LLT label for each AE}

\item{TTTYN}{(character, default = NULL) column name indicating if the
treatemnt was taken}

\item{ARMvar}{(character) column name for treatment group}

\item{seuillab}{(numeric, default = 0.5) seuil vertical d’affichage des
label, permet de ne pas afficher tous les labels pour épurer le graphique
(entrer une valeur de p-value elle sera convertie en -log(10))}

\item{caption}{(boolean, default = TRUE) TRUE/FALSE afficher une note de base
de graphique avec des informations supplémentaires sur le format et le
continu du graphique}

\item{listcol}{(vector of character, default = (“red”,“deepskyblue2”)) liste
de couleurs pour chacun des groupes de traitements dans un vecteur soit avec
les noms implémentés dans R (comme “red”, “blue”…) soit avec des codes
(commençant par # par exemple “#52717F”)}
}
\value{
A Volcano plot
}
\description{
chaque bulle = un EI, occurrence de l’EI pour la taille des
bulles (soit nombre d’EIs, soit nombre de patients ayant eu cet EI). Couleur
des bulles en fonction du groupe de traitement. P-value de Fisher en
ordonnées et pour la saturation des bulles. En abscisse une mesure de
différence entre les deux groupes soit l’absolute risk difference, le risque
ratio, odds ratio, ou incidence rate ratio…
}
\examples{
library(dplyr)
baseEI <- data.frame(idvar = paste0("Patients", round(runif(n = 100, min = 0, max = 100))),
                     Termsvar = round(runif(n = 100, min = 0, max = 20))) \%>\%
  dplyr::mutate(SOCvar = round(Termsvar/10)) \%>\%
  dplyr::mutate(across(everything(), .fns = as.character))

baseTr  <- baseEI \%>\%
  dplyr::select(idvar) \%>\%
  dplyr::distinct() \%>\%
  dplyr::mutate(ARMvar = sample(x = c("Placebo", "Treatment"),
                                size = nrow(.),
                                replace = TRUE),
                TTTYN = sample(x = c("Yes", "No"),
                               size = nrow(.),
                               replace = TRUE,
                               prob = c(0.9, 0.1)))

VolcanoAE(baseEI = baseEI,
          baseTr = baseTr,
          idvar = "idvar",
          Termsvar = "Termsvar",
          TTTYN = "TTTYN",
          ARMvar = "ARMvar")
}
